Fixed debug arrows being drawn in the wrong place
The thin aerofoil debug arrows were being drawn at the position of the centre of pressure for time t-1. This is because I was trying to save time by reusing the stored point of action for the force rather than computing a new one. The confusing issue now though is that the arrows actually represent the forces from the previous time step; not the current one. Not really sure how to fix that, or whether it's worth fixing.