Added Groups and Fixed Lift Calculations
Aero bodies can now be assigned to a group by adding an AeroGroup component to a mutual parent object. This then calculates the total area of the bodies and gives a scaling factor for them based on a desired planform area.
Lift calculation is fixed by properly normalising the lift direction and using planform area instead of profile area for induced drag